added notebook
